Rename Provider to ScoresProvider; add implementation
This renames the Provider as ScoresProvider and
starts a simple implementation: specifically:

 - There is a MatrixCursor called mScores that
   gets some dummy data in onCreate()
 - query() returns mScores

MatrixCursor is a simple 2D table for storing data
and can be passed to anything asking for a Cursor.

This implements the ‘R’ in CRUD; the ‘CRD’ are not
supported.  The one question is whether getType()
needs an implementation; it returns null for now.